requirements.txt
pandas
ipyKernel 
openpyxl
selenium 
python-dotenv